Cleaned code

+++ include/engine/utils/Utilities.h
@@ -8,7 +8,9 @@
 #include "engine\utils\variable\datatype\Vector3.h"
 #include "engine\utils\variable\datatype\Vector4.h"
 #include "engine\utils\variable\datatype\Macros.h"
-#include "engine\utils\Globals.h"
+#include "engine\utils\global\event\GKey.h"
+#include "engine\utils\global\event\GMouse.h"
+#include "engine\utils\global\GScreen.h"
 #include "engine\utils\OpenGL.h"
 #include <sstream>
 #include <stdlib.h>
@@ -19,29 +21,23 @@
 class Math
 {
 public:
-	static GLfloat round2f(GLfloat p_f, Sint32 p_decimalPlaces)
-	{
+	static GLfloat round2f(GLfloat p_f, Sint32 p_decimalPlaces) {
 		return GLfloat(roundf(p_f * GLfloat(pow(10, p_decimalPlaces))) / pow(10, p_decimalPlaces));
 	}
-
-	static GLfloat sind(GLfloat p_f) /* sin of decimal number */
-	{
+	// sin of decimal number
+	static GLfloat sind(GLfloat p_f) {
 		return GLfloat(sin((double)(p_f * 3.14159265359 / 180)));
 	}
-
-	static GLfloat cosd(GLfloat p_f) /* cos of decimal number */
-	{
+	// cos of decimal number
+	static GLfloat cosd(GLfloat p_f)  {
 		return GLfloat(cos((double)(p_f * 3.14159265359 / 180)));
 	}
-
-	static GLfloat tand(GLfloat p_f) /* tan of decimal number */
-	{
+	// tan of decimal number
+	static GLfloat tand(GLfloat p_f) {
 		return GLfloat(tan((double)(p_f * 3.14159265359 / 180)));
 	}
-
 	// Does not use z-axis rotation
-	static Vector3<GLfloat> computeDirection(Vector3<GLfloat> p_rotation)
-	{
+	static Vector3<GLfloat> computeDirection(Vector3<GLfloat> p_rotation) {
 		GLfloat ydist = GLfloat(2 * sind(p_rotation.x));
 		GLfloat xdist = GLfloat(sqrt(pow(2, 2) - pow(ydist, 2)));
 		GLfloat zdist = GLfloat(xdist * sind(p_rotation.y));
@@ -49,8 +45,7 @@ public:
 		return Vector3< GLfloat >(zdist, -ydist, xdist).getNormal();
 	}
 
-	static Vector3<GLfloat> flipOverPoint(Vector3<GLfloat> p_start, Vector3<GLfloat> p_len, Vector3<GLfloat> p_point, bool p_flip[3])
-	{
+	static Vector3<GLfloat> flipOverPoint(Vector3<GLfloat> p_start, Vector3<GLfloat> p_len, Vector3<GLfloat> p_point, bool p_flip[3]) {
 		if(p_flip[0]) p_start.x = (p_start.x - (p_start.x - p_point.x) * 2) - p_len.x;
 		if(p_flip[1]) p_start.y = (p_start.y - (p_start.y - p_point.y) * 2) - p_len.y;
 		if(p_flip[2]) p_start.z = (p_start.z - (p_start.z - p_point.z) * 2) - p_len.z;
@@ -60,8 +55,8 @@ public:
 	static void initPerlin(Uint32 p_seed);
 	static GLfloat perlinNoise(GLfloat x, GLfloat y, GLfloat z, Sint16 p_octaves, GLfloat p_persistence);
 
-	static void castRay3d(Vector3<GLfloat> p_start, Vector3<GLfloat> p_direction, Vector3<GLfloat> p_boxPosition, Vector3<GLfloat> p_boxDimension, GLfloat &p_near, GLfloat &p_far, Sint8 &p_side);
-	static void castBox3d(Vector3<GLfloat> p_start, Vector3<GLfloat> p_dimension, Vector3<GLfloat> p_direction, Vector3<GLfloat> p_boxPosition, Vector3<GLfloat> p_boxDimension, GLfloat &p_near, GLfloat &p_far, Sint8 &p_side);
+	static void castRay3d(Vector3<GLfloat> p_start, Vector3<GLfloat> p_direction, Vector3<GLfloat> p_boxPosition, Vector3<GLfloat> p_boxDimension, GLdouble &p_near, GLdouble &p_far, Sint8 &p_side);
+	static void castBox3d(Vector3<GLfloat> p_start, Vector3<GLfloat> p_dimension, Vector3<GLfloat> p_direction, Vector3<GLfloat> p_boxPosition, Vector3<GLfloat> p_boxDimension, GLdouble &p_near, GLdouble &p_far, Sint8 &p_side);
 private:
 	static GLfloat perlin(GLfloat x, GLfloat y, GLfloat z);
 
@@ -73,10 +68,8 @@ class Util
 {
 public:
 	template<class T>
-	static std::string numToString(T n, Uint16 p_decimalPlaces = 0)
-	{
-		try
-		{
+	static std::string numToString(T n, Uint16 p_decimalPlaces = 0) {
+		try {
 			return std::string(std::to_string(n)).substr(0, std::string(std::to_string(n)).find('.', 0) + p_decimalPlaces + (p_decimalPlaces > 0 ? 1 : 0));
 		}
 		catch(...) {
@@ -86,8 +79,7 @@ public:
 	}
 
 	template<class T>
-	static std::string stringToNum(T s)
-	{
+	static std::string stringToNum(T s) {
 		std::string save = "";
 		std::ostringstream convert;
 		convert.str(std::string());
@@ -101,11 +93,9 @@ public:
 		}
 	}
 
-	static std::string subchar(char* chararray, Uint32 start, Uint32 end)
-	{
+	static std::string subchar(char* chararray, Uint32 start, Uint32 end) {
 		std::string subbed;
-		for(Uint32 i = start; i < end + 1; i++)
-		{
+		for(Uint32 i = start; i < end + 1; i++) {
 			subbed += chararray[i];
 		}
 		return subbed;
